requirements.txt
# Dependências do Projeto Opportunity
# Sistema de coleta e análise de dados de academias

# Análise de dados
pandas>=1.5.0
numpy>=1.21.0
jupyter>=1.0.0
notebook>=6.4.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.0.0

# Banco de dados
sqlite3

# Geocodificação
geopy>=2.2.0
googlemaps>=4.0.0

# Utilitários
python-dotenv>=0.19.0
tqdm>=4.64.0

# Análise e visualização
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Processamento de dados
openpyxl>=3.0.0
xlrd>=2.0.0

# Desenvolvimento e testes
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
